# docs/swagger.yaml
basePath: /
definitions:
...
  models.Role:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.SMSConfirmation:
    properties:
      confirmed:
        type: boolean
      confirmed_at:
        type: string
      document_id:
        type: integer
      id:
        type: integer
      phone:
        type: string
      sent_at:
        type: string
      sms_code:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@company.com
    name: Тур Компани
    url: http://contact.company.com
  description: Aпишка для управления туристической компанией (документация Swagger).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://your-terms.com/terms/
  title: Tour Company API
  version: "1.0"
paths:
...
  /roles:
    get:
      description: Возвращает список всех ролей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список ролей
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Создает новую роль в системе
      parameters:
      - description: Данные новой роли
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать роль
      tags:
      - Roles
  /roles/{id}:
    delete:
      description: Удаляет роль по ID
      parameters:
      - description: ID роли
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить роль
      tags:
      - Roles
    get:
      description: Возвращает данные одной роли
      parameters:
      - description: ID роли
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить роль по ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Обновляет данные роли по ID
      parameters:
      - description: ID роли
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные роли
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить роль
      tags:
      - Roles
  /roles/count:
    get:
      description: Возвращает общее количество ролей в системе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить количество ролей
      tags:
      - Roles
  /roles/with-user-counts:
    get:
      description: Возвращает список ролей с количеством пользователей для каждой
        роли
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить роли с количеством пользователей
      tags:
      - Roles
  /sms/{document_id}:
    delete:
      description: Удаляет все SMS-подтверждения по документу
      parameters:
      - description: ID документа
        in: path
        name: document_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить SMS-подтверждения
      tags:
      - SMS
    get:
      description: Возвращает последнее SMS по документу
      parameters:
      - description: ID документа
        in: path
        name: document_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SMSConfirmation'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить последнее SMS
      tags:
      - SMS
  /sms/confirm:
    post:
      consumes:
      - application/json
      description: Подтверждает введённый код по ID документа
      parameters:
      - description: ID документа и код подтверждения
        in: body
        name: input
        required: true
        schema:
          properties:
            code:
              type: string
            document_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Подтвердить SMS-код
      tags:
      - SMS
  /sms/resend:
    get:
      description: Повторно отправляет SMS по ID документа
      parameters:
      - description: ID документа
        in: query
        name: document_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Повторная отправка SMS
      tags:
      - SMS
  /sms/send:
    post:
      consumes:
      - application/json
      description: Отправляет SMS с кодом подтверждения на указанный номер
      parameters:
      - description: Данные для отправки SMS
        in: body
        name: input
        required: true
        schema:
          properties:
            document_id:
              type: integer
            phone:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Отправить SMS
      tags:
      - SMS
...
swagger: "2.0"